Changed constructor to made tests, but cannot.
OkHttp has some flaws in their API, although it is
widely used. Response and Request do not implement equals,
so we cannot mock it, as when requires to compare to return a
value. This could be solved in more modern versions of Mockito,
but they need a higher java runtime, and rmi is deprecated as Java8.